feat: add commands for managing RETL sources including preview and validation functionality
